refactor: share FX tooltip messages
